undeepfreeze options
